+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+- Improve ```.setGlobalConfiguration()``` api
+    - Accepts method as configuration, the first parameter is the current global configuration
+
 ## 0.8.2 (2016-05-31)
 - Introduce ```babel``` task
     - Use ```babel``` and ```babel:watch``` tasks to build your project